requirements.txt
ollama
chromadb
pdfplumber
langchain
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters
unstructured
unstructered[all-docs]
fastembed
pdfplumber
sentence-transformers
elevenlabs

whisper
pyaudio
wave
